refactor(test): migrate users.ts voice chat helpers to three-tier architecture
Move 3 DB-direct seeder functions and 5 DB-direct assertion functions from
api-test-helpers/users.ts to the appropriate tier files. Add re-exports to
preserve all public APIs so consumer tests need zero changes.

Seeders moved: createTestVoiceChatSession, insertTestVoiceChatSession,
insertTestVoiceChatPreparation
Assertions moved: getPushSubscriptionsByEndpoint, getTestVoiceChatSessionStatus,
getTestVoiceChatSessionHeartbeat, getTestVoiceChatEvents,
getTestVoiceChatPreparation

Closes #9376
